// temp/pms/models/com_ever_edu_pms_terms_dto_res_TermsResDto$DetailOnUser.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { com_ever_edu_pms_terms_dto_res_TermsResDto$TranslationDto } from './com_ever_edu_pms_terms_dto_res_TermsResDto$TranslationDto';
export type com_ever_edu_pms_terms_dto_res_TermsResDto$DetailOnUser = {
    termsId?: number;
    translation?: com_ever_edu_pms_terms_dto_res_TermsResDto$TranslationDto;
};