// temp/pms/models/com_ever_edu_pms_user_dto_req_SendVerifyEmailReqDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type com_ever_edu_pms_user_dto_req_SendVerifyEmailReqDto = {
    /**
     * 이름
     */
    name: string;
    /**
     * 생년월일
     */
    birthday?: string;
    /**
     * Email 주소
     */
    email: string;
};