// temp/pms/models/com_ever_edu_pms_user_dto_req_VerifyPhoneNumberReqDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type com_ever_edu_pms_user_dto_req_VerifyPhoneNumberReqDto = {
    /**
     * 이름
     */
    name: string;
    /**
     * 생년월일
     */
    birthday?: string;
    /**
     * 휴대전화번호
     */
    phoneNumber: string;
    /**
     * 검증 코드
     */
    verificationCode: string;
};